Separate sparql queries for virtuoso
